Add bazel build for songs.
This will make it easier to build songs that changed, vs rebuilding
everything with build_songs.sh.  No midi support yet, but it's
easy/obvious how to bolt it on.

This change also adds the lilypond tool as a binary dependency which
will get downloaded automatically when you build something that needs
it.